Update the XML formats slightly to handle services backed by puppet.
New XML format includes a puppet element to allow us to specify that the
services are configured by puppet.  Also, removed simple2.assy for now.  We only
want to have a single Template-Assembly-Deployable for the first phase
deliverable.  And, bumped fedora version to 14 to match the test image we're
using. Updated audrey.rb to handle the new XML formats and generate a valid job
file.

Signed-off-by:  Greg Blomquist <gblomqui@redhat.com>