Created basic audrey Config Server.
This cut of the  Audrey Config Server (Aeolus Config Server) is a very basic
rest-based ruby (sinatra) web application that associates configuration data
with UUIDs.  GETs retrieve the configuration data and POSTs/PUTs create/update
the configuration data.

At the moment there is no validation of the configuration data.  Nor is there
any cross-pollination of instance configurations.  These are both future
additions.

This first cut is mainly to establish the server and the packaging/deployment.

Signed-off-by: Greg Blomquist <gblomqui@redhat.com>