Tying API version to RPM version in config server.
When the config server RPM is built, the version is propagated down to the
config.ru file.  This seeds the rack "settings" environment with a version
property, which can be used by the rest of the application to validate the
version.

Currently, the version is only used for informational purposes (GET /version),
but it makes it easier to integrate the version into validating requests.

Signed-off-by: Greg Blomquist <gblomqui@redhat.com>