Fix for 401 on HTTP PUT
https://bugzilla.redhat.com/show_bug.cgi?id=760377

The fix for this involves including a bit of a hack in the authentication
process.  The underlying problem appears to be that certain older OAuth libraries
(python-oauth2-1.2.1-3.fc15.noarch) do not include the request body in the part
of the request that's signed with HTTP PUTs.  However, newer OAuth libraries
(python-oauth2-1.5.170-2.fc16.noarch) do include the entire request body in the
signature for HTTP PUTs.

The fix introduced in this patch attempts to authenticate the request first by
including the entire request body in the signature validation.  If that fails,
it attempts to authenticate the request by removing the request body (i.e., only
uses headers) in the signature validation.  If that fails, then the request
cannot be authenticated and a 401 is returned.