Fix for https://bugzilla.redhat.com/show_bug.cgi?id=768425
https://bugzilla.redhat.com/show_bug.cgi?id=768425

This bug shows up every several runs of aeolus-configserver-setup.  The result
is that Puppet fails to leave Apache in a "running" state after running through
the setup.

This fix forces several additional configuration steps to execute _before_
attempting to notify Apache to start.  Namely, the configuration of selinux and
the firewall.