Config Server Setup:  Allow proxy configs to reside alongside conductor
What this commit actually does is pull the config server apache proxy configs
out of the vhost configuration and moves them to a separate include file.
Config server was already pulling in all conf files from
/etc/httpd/conf.d/cs443.  This change allows the Config server apache proxy
configuration to exist in either
  - /etc/httpd/conf.d/cs443, or
  - /etc/httpd/conf.d/aeolus-conductor.d
At the moment, the aeolus-conductor.d apache configuration directory doesn't
exist and is not created by aeolus-configure.  Once we get that change into
aeolus-configure, we can push this change in for config server.