Correct configserver-setup keep_going prompt
Previously, the prompt would not recognize invalid input.  You could
press <Enter> and the setup process would continue.  I've adjusted the
prompt to check whether we are running on a tty first, and to reprompt
when a invalid response is received.  Checking for the presence of a
tty, will allow aeolus-configserver-setup to be called by a
non-interactive script (or deployable).