BZ 751088 (refix): Conductor and configserver on the same box
https://bugzilla.redhat.com/show_bug.cgi?id=751088

This patch changes the aeolus-configserver-setup script.  When the script
detects that conductor is already installed, the script will ensure that the
root context of the aeolus-configserver service is updated to match the root
context of the proxy pass settings in apache.

This change allows the oauth library in the aeolus-configserver service to
correctly decipher the encrypted payload because the request base path will
remain unaltered after the proxy pass.